Refactored testConnectionReadHandler. Finished BufferedQueueManager test

+++ include/fractals/persist/announce_model.h
@@ -13,7 +13,7 @@ namespace fractals::persist {
         int id;
         int torrent_id;
         std::string peer_ip;
-        uint peer_port;
+        uint16_t peer_port;
         time_t announce_time;
         int interval;
         std::optional<int> min_interval;
