+++ src/fractals/network/http/Peer.cpp
@@ -2,7 +2,7 @@
 
 namespace fractals::network::http {
 
-    PeerId::PeerId(std::string ip,uint port) : m_ip(ip),m_port(port) {};
+    PeerId::PeerId(std::string ip,uint16_t port) : m_ip(ip),m_port(port) {};
 
     bool PeerId::operator<(const PeerId &p2) const {
         return m_ip < p2.m_ip || (m_ip == p2.m_ip && m_port < p2.m_port);
